Added tests for the intermediary server
